Refactor: FileController, LocalFileStorage
- FileController에서 ResponseEntity 반환하도록 변경
- LocalFileStorage 에서 dowmload 메서드 이름 변경 후 InputStream 반환하도록 변경
