--- files/csv/backup(1).csv
@@ -1,1 +1,0 @@
-test data
