--- files/csv/backup(2).csv
@@ -1,1 +1,0 @@
-test data
