--- files/csv/backup.csv
@@ -1,1 +1,0 @@
-test data
